+++ fili-core/src/main/java/com/yahoo/bard/webservice/data/dimension/datatype/DateDimensionDatatype.java
@@ -2,24 +2,17 @@
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.data.dimension.datatype;
 
-import java.util.Locale;
-
 /**
- * Default set of dimension datatypes.
+ * Date dimension datatype.
  */
-public enum DefaultDimensionDatatype implements DimensionDatatype {
-    TEXT,
-    NUMBER,
-    DATE
-    ;
-
+public class DateDimensionDatatype implements DimensionDatatype {
     @Override
     public String getName() {
-        return name().toLowerCase(Locale.ENGLISH);
+        return "Date";
     }
 
     @Override
     public boolean isA(DimensionDatatype otherType) {
-        return otherType == this;
+        return otherType instanceof DateDimensionDatatype;
     }
 }